Attempt to move contrib.named to contrib.autoname
