+++ examples/contrib/autoname/mixture.py
