+++ pyro/contrib/__init__.py
@@ -8,10 +8,10 @@ Contributed Code
 
 from __future__ import absolute_import, division, print_function
 
-from pyro.contrib import autoguide, gp, named
+from pyro.contrib import autoguide, autoname, gp
 
 __all__ = [
     "autoguide",
+    "autoname",
     "gp",
-    "named",
 ]
