+++ pyro/contrib/autoname/named.py
