+++ tests/doctest_fixtures.py
@@ -4,7 +4,7 @@ import torch
 
 import pyro
 import pyro.contrib.gp as gp
-import pyro.contrib.named as named
+import pyro.contrib.autoname.named as named
 import pyro.distributions as dist
 import pyro.poutine as poutine
 from pyro.infer import EmpiricalMarginal
