+++ tests/test_examples.py
@@ -16,8 +16,8 @@ CPU_EXAMPLES = [
     ['baseball.py', '--num-samples=200', '--warmup-steps=100'],
     ['bayesian_regression.py', '--num-epochs=1'],
     ['contrib/gp/sv-dkl.py', '--epochs=1', '--num-inducing=4'],
-    ['contrib/named/mixture.py', '--num-epochs=1'],
-    ['contrib/named/tree_data.py', '--num-epochs=1'],
+    ['contrib/autoname/mixture.py', '--num-epochs=1'],
+    ['contrib/autoname/tree_data.py', '--num-epochs=1'],
     ['dmm/dmm.py', '--num-epochs=1'],
     ['dmm/dmm.py', '--num-epochs=1', '--num-iafs=1'],
     ['inclined_plane.py', '--num-samples=1'],
